# Conflicts:
#	Assets/levels/cur_seed.json

+++ source/ECS/Components.cpp
@@ -54,7 +54,7 @@ void Components::Card_Storage::reset()
 Components::Input::Input(u8 type, bool hover,
 	std::function<void()> onClick,
 	std::function<void()> onHover,
-	std::function<void()> offHover, s8 z, bool drag, bool col, 
+	std::function<void()> offHover, s32 z, bool drag, bool col, 
 						 std::function<void()> onDrag, 
 						 std::function<void()> offDrag)
 	: type(type), hover(hover), onClick(onClick),
@@ -62,7 +62,7 @@ Components::Input::Input(u8 type, bool hover,
 {
 }
 
-Components::Text::Text(std::string text, s8 fontID, s8 z) : text{ text }, fontID{ fontID }, z{ z } {}
+Components::Text::Text(std::string text, s8 fontID, s32 z) : text{ text }, fontID{ fontID }, z{ z } {}
